Use window function to get top rank in book search
This is a tentative attempt to reduce the multiple queries needed to
retrieve the top-ranked search result in the book search to a single
query. It uses a window function to rank the results by their search
rank, and then take top-ranked results.

It is tentative because I have somewhere used to subqueries, which
seems excessive. I'm not familiar with the Django ORM's window
function syntax. I'll need to investigate a bit further to clean this up.